adding truly modular interface components. try modularpanels.html
demonstrates restrained resizable modules with gridsnapping for size and position. modules are restrained into the panel the reside in. we should use this feature on certain sections where lots of reports and information are needed, as well as certain workflow areas. the main places are the "home" pages for each section. These will list off the general information for each client, and once we get the jquery crud forms done the isps/comps for the office will be a breeze.